% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{CheckWGCNAName}
\alias{CheckWGCNAName}
\title{CheckWGCNAName}
\usage{
CheckWGCNAName(seurat_obj, wgcna_name)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}
}
\description{
CheckWGCNAName
}
\keyword{scRNA-seq}
