% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetActiveWGCNA}
\alias{GetActiveWGCNA}
\title{GetActiveWGCNA}
\usage{
GetActiveWGCNA(seurat_obj)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}
}
\description{
GetActiveWGCNA
}
\keyword{scRNA-seq}
