% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMultiExpr}
\alias{GetMultiExpr}
\title{GetMultiExpr}
\usage{
GetMultiExpr(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}
}
\description{
此函数从元细胞对象获取表达矩阵。
}
\keyword{scRNA-seq}
