% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetWGCNAGenes}
\alias{GetWGCNAGenes}
\title{GetWGCNAGenes}
\usage{
GetWGCNAGenes(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{一个 Seurat}

\item{wgcna_name}{seurat_obj@misc 插槽中 hdWGCNA 实验的名称}
}
\description{
GetWGCNAGenes
}
\keyword{scRNA-seq}
