% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_module_networks}
\alias{plot_module_networks}
\title{Plot Module Networks and UMAPs}
\usage{
plot_module_networks(seurat_obj, plot_path = "./plots/")
}
\arguments{
\item{seurat_obj}{A Seurat object with WGCNA results.}

\item{plot_path}{The directory path to save the output plots.}
}
\description{
This function generates and plots various network visualizations and UMAPs for WGCNA modules.
}
