% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sgwgcna_pipeline}
\alias{run_sgwgcna_pipeline}
\title{Run sgWGCNA Pipeline}
\usage{
run_sgwgcna_pipeline(
  seurat_obj,
  gene_select = "fraction",
  fraction = 0.05,
  wgcna_name = "ASC_consensus",
  metacell_group.by = c("celltype", "sample"),
  metacell_ident.group1 = "celltype",
  metacell_ident.group2 = "sample",
  k = 25,
  max_shared = 12,
  min_cells = 50,
  target_metacells = 250,
  reduction = "harmony",
  dat_expr_group_name = "B cell",
  network_type = "signed",
  n_hubs = 20,
  n_genes_score = 25,
  assay = NULL,
  meta_yazu = "sample",
  dot_plot_group = "sample",
  plot_path = "./plots/",
  plot_height = 10,
  plot_width = 10,
  cols = 6
)
}
\arguments{
\item{seurat_obj}{The input Seurat object.}

\item{gene_select}{The method for gene selection. Can be "fraction" or "variable".}

\item{fraction}{The fraction of cells a gene must be expressed in to be included.}

\item{wgcna_name}{The name for the WGCNA analysis stored in the Seurat object.}

\item{metacell_group.by}{The metadata columns to group cells for metacell analysis.}

\item{metacell_ident.group1}{The primary metadata column for cell identity.}

\item{metacell_ident.group2}{The secondary metadata column for grouping (e.g., sample).}

\item{k}{The number of nearest neighbors for metacell construction.}

\item{max_shared}{The maximum shared neighbors for metacell construction.}

\item{min_cells}{The minimum number of cells required for a group to be included.}

\item{target_metacells}{The target number of metacells to create.}

\item{reduction}{The dimensionality reduction to use for metacell construction.}

\item{dat_expr_group_name}{The specific group name from \code{metacell_ident.group1} to analyze.}

\item{network_type}{The type of network to construct (e.g., "signed").}

\item{n_hubs}{The number of hub genes to identify per module.}

\item{n_genes_score}{The number of genes to use for module scoring.}

\item{assay}{The assay to use for expression data.}

\item{meta_yazu}{The metadata column for grouping in radar plots.}

\item{dot_plot_group}{The metadata column to use for the dot plot.}

\item{plot_path}{The directory path to save the output plots.}

\item{plot_height}{The height of the saved plots in inches.}

\item{plot_width}{The width of the saved plots in inches.}

\item{cols}{The number of columns for arranging feature plots.}
}
\value{
The processed Seurat object with WGCNA results.
}
\description{
A function to encapsulate the hdWGCNA analysis pipeline for a single group.
}
\examples{
\dontrun{
seurat_obj <- run_sgwgcna_pipeline(
  seurat_obj,
  dat_expr_group_name = "CD4 T-cells",
  min_cells = 15,
  plot_path = "wgcna_plots_cd4/"
)
}
}
\keyword{WGCNA}
\keyword{scRNA-seq}
