% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectNetworkGenes.R
\name{SelectNetworkGenes}
\alias{SelectNetworkGenes}
\title{SelectNetworkGenes}
\usage{
SelectNetworkGenes(
  seurat_obj,
  gene_select = "variable",
  fraction = 0.05,
  group.by = NULL,
  gene_list = NULL,
  assay = NULL,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}

\item{gene_select}{如何选择基因？选择 "variable"（变量）、"fraction"（比例）、"all"（全部）或 "custom"（自定义）。}

\item{fraction}{一个数值，决定了一个基因必须在多少最小细胞数中表达才能被包含。例如，fraction = 0.05 意味着一个基因必须在 5\% 的细胞中表达（计数 > 0）才能被包含。}

\item{group.by}{Seurat 对象中的一个元数据列，用于对细胞进行分组（例如 seurat_clusters）。}

\item{gene_list}{一个基因名称的字符向量，仅在 gene_select = "custom" 时使用。}

\item{assay}{seurat_obj 中的 Assay，用于计算模块特征基因。默认为 DefaultAssay(seurat_obj)。}

\item{wgcna_name}{WGCNA 实验的名称。}
}
\description{
选择将用于共表达网络分析的基因
}
\details{
SelectNetworkGenes 允许我们指定将用于共表达网络分析的基因。
这个函数由 SetupForWGCNA 调用。默认情况下，使用 VariableFeatures(seurat_obj) 中的可变特征。
也可以通过 gene_list 参数和设置 gene_select="custom" 来使用自定义基因列表。

我们还可以通过设置 gene_select='fraction' 来识别在数据集中特定比例的细胞中表达量大于 0 的基因。
例如，通过设置 fraction=0.1 和 group.by='seurat_clusters'，此函数将识别出
在至少一个聚类中 10\% 的细胞中表达的基因集合。
}
