% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetActiveWGCNA}
\alias{SetActiveWGCNA}
\title{SetActiveWGCNA}
\usage{
SetActiveWGCNA(seurat_obj, wgcna_name)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}

\item{wgcna_name}{seurat_obj@misc 插槽中 hdWGCNA 实验的名称}
}
\description{
SetActiveWGCNA
}
\keyword{scRNA-seq}
