% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetDatExpr}
\alias{SetDatExpr}
\title{SetDatExpr}
\usage{
SetDatExpr(
  seurat_obj,
  group_name,
  use_metacells = TRUE,
  group.by = NULL,
  multi.group.by = NULL,
  multi_group_name = NULL,
  return_seurat = TRUE,
  assay = NULL,
  slot = "data",
  layer = "data",
  mat = NULL,
  features = NULL,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}

\item{group_name}{一个字符串，包含提供的 group.by 列中或 Seurat Idents 中的一个组。可以提供一个字符向量以一次选择多个组。}

\item{use_metacells}{一个逻辑值，确定我们是使用元细胞 (TRUE) 还是完整的表达矩阵 (FALSE)}

\item{group.by}{一个字符串，包含 Seurat 对象中具有细胞组（簇、细胞类型等）的列的名称。如果为 NULL（默认值），hdWGCNA 使用 Seurat Idents 作为组。}

\item{multi.group.by}{一个字符串，包含 Seurat 对象中具有用于共识 WGCNA 的组（数据集、样本、条件等）的列的名称}

\item{multi_group_name}{一个字符串，包含 multi.group.by 列中存在的组的名称。}

\item{assay}{Seurat 对象中分析的名称}

\item{slot}{用于提取聚合数据的插槽。默认 = 'counts'。插槽与 Seurat v4 一起使用，而不是层。}

\item{layer}{用于提取聚合数据的层。默认 = 'counts'。层与 Seurat v5 一起使用，而不是插槽。}

\item{mat}{一个包含基因表达数据的矩阵。使用此参数提供矩阵会忽略其他选项。这几乎专门用于伪批量分析。}

\item{features}{要用于覆盖先前设置的特征的特征列表。}

\item{wgcna_name}{一个字符串，包含 seurat_obj@misc 中 WGCNA 插槽的名称。默认 = NULL，它检索当前活动的 WGCNA 数据}
}
\description{
此函数指定用于共表达网络分析的基因表达矩阵。
}
\details{
SetDatExpr 是 hdWGCNA 管道的一个关键函数，它确定将用于网络分析的基因表达矩阵。
我们通常使用此函数使用 group.by 参数选择一个细胞类型或一组细胞类型进行网络分析，但我们提供
用于进一步定制的其他参数。
}
\keyword{scRNA-seq}
