% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMetacellObject}
\alias{SetMetacellObject}
\title{SetMetacellObject}
\usage{
SetMetacellObject(seurat_obj, metacell_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}

\item{metacell_obj}{元细胞 Seurat 对象}

\item{wgcna_name}{seurat_obj@misc 插槽中 hdWGCNA 实验的名称}
}
\description{
SetMetacellObject
}
\keyword{scRNA-seq}
