% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMultiExpr}
\alias{SetMultiExpr}
\title{SetMultiExpr}
\usage{
SetMultiExpr(
  seurat_obj,
  group_name,
  use_metacells = TRUE,
  group.by = NULL,
  multi.group.by = NULL,
  multi_groups = NULL,
  assay = NULL,
  slot = "data",
  layer = "data",
  mat = NULL,
  mat_group_delim = 3,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{一个 Seurat 对象}

\item{group_name}{一个字符串，包含提供的 group.by 列中或 Seurat Idents 中的一个组。}

\item{use_metacells}{一个逻辑值，确定我们是使用元细胞 (TRUE) 还是完整的表达矩阵 (FALSE)}

\item{group.by}{一个字符串，包含 Seurat 对象中具有细胞组（簇、细胞类型等）的列的名称。如果为 NULL（默认值），hdWGCNA 使用 Seurat Idents 作为组。}

\item{multi.group.by}{一个字符串，包含 Seurat 对象中具有用于共识 WGCNA 的组（数据集、样本、条件等）的列的名称}

\item{multi_groups}{一个字符向量，包含要选择的组的名称}

\item{assay}{Seurat 对象中分析的名称}

\item{slot}{Seurat 对象中插槽的名称（counts、data）}

\item{layer}{用于提取聚合数据的层。默认 = 'counts'。层与 Seurat v5 一起使用，而不是插槽。}

\item{mat}{一个包含基因表达数据的矩阵。使用此参数提供矩阵会忽略其他选项。这几乎专门用于伪批量分析。}

\item{wgcna_name}{一个字符串，包含 seurat_obj@misc 中 WGCNA 插槽的名称。默认 = NULL，它检索当前活动的 WGCNA 数据}
}
\description{
此函数根据元细胞表达矩阵、完整表达矩阵或提供的伪批量表达矩阵，为共识 WGCNA 设置表达矩阵输入。
}
\keyword{scRNA-seq}
